refactor: move add to list to component
